// Knowit.Umbraco.Bellissima.InstantBlockPreview.Frontend/src/api/types.gen.ts
export type DataTypePropertyPresentationModel = {
    alias: string;
    value?: unknown;
};

export type DataTypePropertyReferenceModel = {
    name: string;
    alias: string;
};

export type DataTypeReferenceResponseModel = {
    contentType: (DataTypeContentTypeReferenceModel);
    properties: Array<(DataTypePropertyReferenceModel)>;
};

export type DataTypeResponseModel = {
    name: string;
    editorAlias: string;
    editorUiAlias: string;
    values: Array<(DataTypePropertyPresentationModel)>;
    id: string;
    isDeletable: boolean;
    canIgnoreStartNodes: boolean;
};